[feature] user rate

+++ workspace/web_api/persistence/src/main/java/com/notrika/gympin/persistence/entity/rating/UserRate.java
@@ -29,6 +29,6 @@ public class UserRate extends BaseEntity {
     private User judgingUser;
 
     @Column(name = "rate")
-    private Integer rate;
+    private Float rate;
 
 }
